Added the optional top 3 book function.
This isn't in my video but an explanation of how it works is. It prints top 3 viewed books. Then using the key lambda it sorts all the books in the list to where they are in order of value of checkouts. It then for every book in the top three it prints it.